continued mpc formulation

+++ stepping_stone.cpp
@@ -1,7 +1,7 @@
 #include "stepping_stone.h"
 
 
-SteppingStone::SteppingStone(Eigen::Vector3d point, Eigen::Vector3d normal, float width, float height) 
+SteppingStone::SteppingStone(Eigen::Vector3d point, Eigen::Vector3d normal, double width, double height) 
 {
     this->width = width;
     this->height = height;
